Add comments. Minor refactoring according to PR comments
