+++ GameOfRunes/GameOfRunes/Custom Nodes/ButtonNode/ButtonNodeResponderType.swift
@@ -1,0 +1,18 @@
+//
+//  ButtonNodeResponderType.swift
+//  GameOfRunes
+//
+//  Created by Brian Yen on 11/3/20.
+//  Copyright © 2020 TeamHoWan. All rights reserved.
+//
+
+import SpriteKit
+
+/**
+ A type that can respond to `ButtonNode` button press events.
+ This should be implemented by the `SKScene` that the `ButtonNode`
+ is added to.
+ */
+protocol ButtonNodeResponderType : class {
+    func buttonPressed(button: ButtonNode)
+}
