+++ GameOfRunes/GameOfRunes/Entities/EntityManager.swift
@@ -9,6 +9,9 @@
 import SpriteKit
 import GameplayKit
 
+/**
+ Class to manage all Entities within the game.
+ */
 class EntityManager {
     lazy var componentSystems: [GKComponentSystem] = {
         let manaSystem = GKComponentSystem(componentClass: ManaComponent.self)
@@ -77,7 +80,7 @@ class EntityManager {
         // Player Loses the Game
         if let playerHealthEntity = playerHealthEntity,
             let playerHealthComponent = playerHealthEntity.component(ofType: HealthComponent.self),
-            playerHealthComponent.healthPoints == 0,
+            playerHealthComponent.healthPoints <= 0,
             let gameStateMachine = gameStateMachine,
             let gameEndState = gameStateMachine.state(forClass: GameEndState.self) {
             gameEndState.didWin = false
@@ -99,23 +102,30 @@ class EntityManager {
         }
         add(enemyEntity)
     }
-    
+
+    /** Gets all entities of a particular `Team`. */
     func entities(for team: Team) -> [GKEntity] {
         entities.compactMap { $0.component(ofType: TeamComponent.self)?.team == team ? $0 : nil }
     }
-    
+
+    /** Gets all the `MoveComponents` of entities in a particular `Team` */
     func moveComponents(for team: Team) -> [MoveComponent] {
         let entitiesToMove = entities(for: team)
         return entitiesToMove.compactMap { $0.component(ofType: MoveComponent.self) }
     }
 
+    /** Decrements the Player's health by 1 point. */
     func decreasePlayerHealth() {
         if let playerHealthEntity = playerHealthEntity,
             let playerHealthComponent = playerHealthEntity.component(ofType: HealthComponent.self) {
             playerHealthComponent.healthPoints -= 1
         }
     }
-    
+
+    /**
+     Removes all `EnemyEntity`s whose `GestureComponent` corresponds
+     to `gesture`.
+     */
     func removeMonstersWithGesture(gesture: CustomGesture) {
         for enemyEntity in entities(for: .enemy) {
             guard let gestureComponent = enemyEntity.component(ofType: GestureComponent.self) else {
